requirements.txt
Django==4.2.8
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
openai>=1.52.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
flask==3.0.0
gunicorn
